build(backoffice): npm build and infra components
