+++ src/contexts/backoffice/commerces/domain/InvalidPriceError.ts
@@ -1,0 +1,1 @@
+export default class InvalidPriceError extends Error {}
