+++ src/contexts/backoffice/commerces/domain/ProductWithPrice.ts
@@ -1,5 +1,5 @@
 import Product from "@backoffice-contexts/products/domain/Product";
-import PriceVo from "@pms-contexts/products/domain/PriceVo";
+import PriceVo from "@backoffice-contexts/commerces/domain/PriceVo";
 
 export default interface ProductWithPrice {
     product: Product,
